% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{calc_acc}
\alias{calc_acc}
\title{Calculate the accuracy (FDR, Power, F1 score) if truth is available}
\usage{
calc_acc(est, truth)
}
\arguments{
\item{est}{the selected relavent features}

\item{truth}{A vector of the true relevant features}
}
\value{
a vector of size three: fdr, power, and F1 score
}
\description{
Calculate the accuracy (FDR, Power, F1 score) if truth is available
}
