% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{calc_inc_rate}
\alias{calc_inc_rate}
\title{Calculate the inclusion rate}
\usage{
calc_inc_rate(mss, q = 0.05, sum_in_denom = TRUE)
}
\arguments{
\item{mss}{a list of the selection of multiple data splitting}

\item{q}{the nominal FDR level}

\item{sum_in_denom}{use the modified inclusion rate (\code{TRUE}) or the original inclusion rate (\code{FALSE})}
}
\value{
the inclusion rate for each feature
}
\description{
Calculate the inclusion rate
}
