% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{calc_tau}
\alias{calc_tau}
\title{Calculate the cutoff the mirror statistics}
\usage{
calc_tau(ms, q = 0.05)
}
\arguments{
\item{ms}{Array of mirror statistics}

\item{q}{nominal FDR level (default: 0.05)}
}
\value{
a cutoff
}
\description{
Calculate the cutoff the mirror statistics
}
