% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_diff}
\alias{cluster_diff}
\title{Calculate the difference across two clusters for each feature}
\usage{
cluster_diff(x, cl)
}
\arguments{
\item{x}{the data matrix}

\item{cl}{the clustering label}
}
\value{
a vector of differences
}
\description{
Calculate the difference across two clusters for each feature
}
