% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{dd}
\alias{dd}
\title{Wrapper for the naive double-dipping method}
\usage{
dd(sce, params, test.use = "t", q = 0.05)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{params}{parameters passed to the clustering step \code{perform_clustering}}

\item{test.use}{The hypothesis testing for DE test. Possible choices: \code{t}, \code{wilcox}, \code{poisson}, and \code{negbinom}}

\item{q}{nominal FDR level}
}
\value{
a vector of selected significant features
}
\description{
Wrapper for the naive double-dipping method
}
