% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{debias_symmetry}
\alias{debias_symmetry}
\title{Debias the statistics under the null for symmetry}
\usage{
debias_symmetry(t, prop = 0.5)
}
\arguments{
\item{t}{the statistics to be debiased}

\item{prop}{the proportion used to estimate the mean/median of statistics among the null features}
}
\value{
the debiased statistics
}
\description{
Debias the statistics under the null for symmetry
}
