% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{ds.matrix}
\alias{ds.matrix}
\title{DS procedure for Matrix}
\usage{
\method{ds}{matrix}(x, kmeans_whiten = FALSE, Sigma = NULL, q = 0.05, debias = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{kmeans_whiten}{whether to perform the whitening procedure for kmean}

\item{Sigma}{the covariance matrix, if null, estimate from the data}

\item{q}{the nominal FDR level}

\item{debias}{whether to debias (useful when correlation is high)}
}
\value{
a list with the selected set, the mirror statistics, two signal measurements for two halves
}
\description{
DS procedure for Matrix
}
