% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{ds.SingleCellExperiment}
\alias{ds.SingleCellExperiment}
\title{DS procedure for Seurat object}
\usage{
\method{ds}{SingleCellExperiment}(
  sce,
  params1 = NULL,
  params2 = NULL,
  seed = NA,
  test.use = "t",
  signal_measurement = "tstat",
  test.from.raw = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{params1}{list of parameters to be passed to \code{perform_clustering} for the first half of the data splitting}

\item{params2}{list of parameters to be passed to \code{perform_clustering} for the second half of the data splitting.}

\item{seed}{random seed for reproducibility}

\item{test.use}{The hypothesis testing for DE test. Possible choices: \code{t}, \code{wilcox}, \code{poisson}, and \code{negbinom}}

\item{signal_measurement}{the signal measurement. Possible choices: the test statistic \code{tstat}, or the signed p-value \code{pval}}

\item{test.from.raw}{whether to perform the testing on the original data without normalization or on the normalized data}
}
\value{
a vector of mirror statistics
}
\description{
DS procedure for Seurat object
}
