% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est.Sigma}
\alias{est.Sigma}
\title{estimate the covariance matrix (assuming there are two clusters)}
\usage{
est.Sigma(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
estimate covariance matrix
}
\description{
estimate the covariance matrix (assuming there are two clusters)
}
