% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{gen_data_normal}
\alias{gen_data_normal}
\title{Generate simulation data with two gaussians}
\usage{
gen_data_normal(
  n = 100,
  p = 200,
  delta = 3,
  prop = 0.1,
  rho = 0,
  prop_cl1 = 0.5
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of features}

\item{delta}{signal strength}

\item{prop}{proportion of non-null features}

\item{rho}{correlation parameter in the covariance matrix}

\item{prop_cl1}{proportion of the first cluster}
}
\description{
Generate simulation data with two gaussians
}
