% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{gen_data_pois.matrix}
\alias{gen_data_pois.matrix}
\title{Matrix method for gen_data_pois (S3)}
\usage{
gen_data_pois.matrix(Lambda, rho = 0.5, block_size = 10)
}
\arguments{
\item{Lambda}{parameter matrix}

\item{rho}{correlation parameter}

\item{block_size}{the copula correlation is set up in each block (mainly for speed up copula for multivariate distribution)}
}
\value{
data matrix of the same size of \code{Lambda}
}
\description{
Matrix method for gen_data_pois (S3)
}
