% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{gen_data_pois}
\alias{gen_data_pois}
\title{Generate Poisson data with latent structure}
\usage{
gen_data_pois(
  n,
  p,
  delta = 3,
  prop_imp = 0.1,
  rho = 0.5,
  block_size = 10,
  type = "discrete",
  sigma = 0
)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of features}

\item{delta}{Effect size parameter}

\item{prop_imp}{Proportion of important features (default: 0.1)}

\item{rho}{AR(1) correlation parameter (default: 0.5)}

\item{block_size}{Size of correlation blocks (default: 10)}

\item{type}{Type of latent variable: "discrete" (binary) or "continuous" (default: "discrete")}

\item{sigma}{Additional noise standard deviation (default: 0)}
}
\value{
List containing X (data matrix) and L (latent variable)
}
\description{
Generate Poisson data with latent structure
}
