% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{mds.matrix}
\alias{mds.matrix}
\title{MDS procedure for matrix}
\usage{
\method{mds}{matrix}(
  x,
  M = 10,
  q = 0.05,
  ncores = 1,
  tied.method = "fair",
  sum_in_denom = TRUE,
  ...
)
}
\arguments{
\item{x}{data matrix}

\item{M}{number of repetitions}

\item{q}{nominal FDR level}

\item{ncores}{number of cores}

\item{tied.method}{handling the tied values when determing the cutoff}

\item{sum_in_denom}{way for calculating the inclusion rate}

\item{...}{optional arguments passed to \code{ds.matrix}}
}
\value{
a list of mirror statistics
}
\description{
MDS procedure for matrix
}
