% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{mds1}
\alias{mds1}
\title{Conduct multiple data splitting}
\usage{
mds1(sce, M = 2, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{M}{The number of data splitting}

\item{...}{Arguments passed to \code{ds}}
}
\value{
a list of the mirror statistics from multiple data splitting
}
\description{
Step 1 of \code{mds}
}
