% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{mds1_parallel}
\alias{mds1_parallel}
\title{Conduct multiple data splitting in parallel}
\usage{
mds1_parallel(sce, M = 2, ncores = 10, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{M}{The number of data splitting}

\item{ncores}{The number of cores used for parallel computing. It is better to be a factor of \code{M}.}

\item{...}{Arguments passed to \code{ds}}
}
\value{
a list of the mirror statistics from multiple data splitting
}
\description{
Step 1 (in parallel) of \code{mds}
}
