% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{mds2}
\alias{mds2}
\title{Aggregate multiple data splitting results, and return a selection set}
\usage{
mds2(mss, q = 0.05, tied.method = "fair", sum_in_denom = TRUE)
}
\arguments{
\item{mss}{A list, which contains the selections from multiple data splitting procedures}

\item{q}{The nominal FDR level (default: 0.05)}

\item{tied.method}{The method for handling the tied values.}

\item{sum_in_denom}{when calculating the inclusion rate, use the modified inclusion rate (\code{TRUE}) or the original inclusion rate (\code{FALSE})}
}
\value{
the selected relevant features
}
\description{
Aggregate multiple data splitting results, and return a selection set
}
