% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mirror_stat}
\alias{mirror_stat}
\title{Calculate the mirror statistics}
\usage{
mirror_stat(d1, d2)
}
\arguments{
\item{d1}{signal measurement from the first half}

\item{d2}{signal measurement from the second half}
}
\value{
a vector of mirror statistics
}
\description{
Calculate the mirror statistics
}
