% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forked_seurat.R
\name{myDiffTTest}
\alias{myDiffTTest}
\title{Modified Seurat::DiffTTest by extending the output with statistics in addition to p-values}
\usage{
myDiffTTest(data.use, cells.1, cells.2, verbose = TRUE)
}
\arguments{
\item{data.use}{Data matrix to test}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{verbose}{Print a progress bar}
}
\description{
Differential expression testing using Student's t-test
}
\details{
Identify differentially expressed genes between two groups of cells using the Student's t-test
}
