% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forked_seurat.R
\name{myFindMarkers}
\alias{myFindMarkers}
\title{Gene expression markers of identity classes}
\usage{
myFindMarkers(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods and to specific DE methods}
}
\description{
Wrapper for \code{Seurat::FindMarkers} that uses modified (extended) tests
}
\details{
Finds markers (differentially expressed genes) for identity classes
}
