% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forked_seurat.R
\name{myGLMDETest}
\alias{myGLMDETest}
\title{Modified Seurat::GLMDETest by extending the output with statistics in addition to p-values}
\usage{
myGLMDETest(
  data.use,
  cells.1,
  cells.2,
  min.cells = 3,
  latent.vars = NULL,
  test.use = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data.use}{Data to test}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{min.cells}{Minimum number of cells threshold}

\item{latent.vars}{Latent variables to test}

\item{test.use}{parameterizes the glm}

\item{verbose}{Print progress bar}
}
\value{
Returns a p-value ranked matrix of putative differentially expressed
}
\description{
Modified Seurat::GLMDETest by extending the output with statistics in addition to p-values
}
