% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{perform_clustering}
\alias{perform_clustering}
\title{Perform Clustering on a SingleCellExperiment object}
\usage{
perform_clustering(
  sce,
  ret_obj = FALSE,
  normalized_method = "LogNormalize",
  use.kmeans = FALSE,
  kmeans.nstart = 1,
  resolution = 0.5,
  step = 0.05,
  maxiter = 40,
  louvain.nstart = 1,
  louvain.alg = 1,
  seed.cluster = 0,
  kmeans.whiten = FALSE,
  pca.whiten = FALSE,
  npcs = 10,
  k.param = 20,
  test.use = "t",
  test.from.raw = FALSE,
  signal_measurement = "tstat",
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{ret_obj}{(Default: \code{FALSE}) whether to return the \code{sce} after operation instead of returning the signal measurement}

\item{normalized_method}{Normalization method. Possible choices: \code{LogNormalize}, \code{sct}, or \code{none}}

\item{use.kmeans}{if \code{TRUE}, then clustering using kmeans with two clusters, otherwise, use \code{Seurat::FindClusters} and need to find \code{resolution} to achieve two clusters}

\item{kmeans.nstart}{the \code{nstart} parameter for \code{kmeans}}

\item{resolution}{the \code{resolution} parameter for \code{Seurat::FindClusters}}

\item{step}{To find a resolution with desired number of clusters, the step size for changing the \code{resolution} parameter}

\item{maxiter}{maximum iteration for searching \code{resolution} to have two clusters}

\item{louvain.nstart}{the \code{n.start} parameter for \code{Seurat::FindClusters}}

\item{louvain.alg}{the \code{algorithm} parameter for \code{Seurat::FindClusters}}

\item{seed.cluster}{the \code{random.seed} parameter for \code{Seurat::FindClusters}}

\item{kmeans.whiten}{whether to whitening for the kmeans clustering}

\item{pca.whiten}{whether to whitening for PCA if using Louvain algorithms}

\item{npcs}{number of PCs, the parameter \code{npcs} used in \code{Seurat::RunPCA}}

\item{k.param}{the parameter \code{k.param} used in \code{Seurat::FindNeighbors}}

\item{test.use}{The hypothesis testing for DE test. Possible choices: \code{t}, \code{wilcox}, \code{poisson}, and \code{negbinom}}

\item{test.from.raw}{whether to perform the testing on the original data without normalization or on the normalized data}

\item{signal_measurement}{the signal measurement. Possible choices: the test statistic \code{tstat}, or the signed p-value \code{pval}}

\item{verbose}{whether to print internal log messages}
}
\value{
a vector of signal measurements for each feature
}
\description{
Perform Clustering on a SingleCellExperiment object
}
