% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forked_seurat.R
\name{rankSumTestWithCorrelation}
\alias{rankSumTestWithCorrelation}
\title{Wilcoxon rank sum test (adapted from \code{limma::rankSumTestWithCorrelation})}
\usage{
rankSumTestWithCorrelation(index, statistics, correlation = 0, df = Inf)
}
\arguments{
\item{index}{any index vector such that \code{statistics[index]} contains the values of the statistic for the test group.}

\item{statistics}{numeric vector giving values of the test statistic.}

\item{correlation}{numeric scalar, average correlation between cases in the test group.  Cases in the second group are assumed independent of each other and other the first group.}

\item{df}{degrees of freedom which the correlation has been estimated.}
}
\value{
Numeric vector of length 4 containing the \code{left.tail} and \code{right.tail} p-values and the corresponding test statistics.
}
\description{
Wilcoxon rank sum test (adapted from \code{limma::rankSumTestWithCorrelation})
}
