% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cluster_test.R
\name{sel_inc_rate}
\alias{sel_inc_rate}
\title{Perform selection based on inclusion rate}
\usage{
sel_inc_rate(inc_rate, q = 0.05, tied.method = "fair", tol = 1e-10)
}
\arguments{
\item{inc_rate}{inclusion rate}

\item{q}{nominal FDR level}

\item{tied.method}{the method for handling the tied values. Possible choices: \describe{
\item{\code{include}}{include the tied rate}
\item{\code{exclude}}{exclude the tied rate}
\item{\code{fair} (default)}{a compromise way which compare the number of tied values before and after the cutoff}
}}

\item{tol}{tolerance when comparing the equality of two doubles}
}
\value{
the selected relevant features
}
\description{
Perform selection based on inclusion rate
}
