% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata_1ct}
\alias{simdata_1ct}
\title{Datasets
Demo synthetic scRNA-seq data with one cell type based on \code{DuoClustering2018::sce_full_Zhengmix4eq()}}
\format{
a list for \code{simdata_1ct}
}
\usage{
simdata_1ct
}
\description{
Datasets
Demo synthetic scRNA-seq data with one cell type based on \code{DuoClustering2018::sce_full_Zhengmix4eq()}
}
\keyword{datasets}
