% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata_2ct}
\alias{simdata_2ct}
\title{Demo synthetic scRNA-seq data with two cell types based on \code{DuoClustering2018::sce_full_Zhengmix4eq()}}
\format{
a list for \code{simdata_2ct}
}
\usage{
simdata_2ct
}
\description{
Demo synthetic scRNA-seq data with two cell types based on \code{DuoClustering2018::sce_full_Zhengmix4eq()}
}
\keyword{datasets}
